Move daml-assistant logic out of daml-project-config.
